+++ doc/version_history.rst
@@ -11,6 +11,7 @@ v1.19.0
 
 * Add the `rules.MTM2OutClosedLoopControl` rule.
 * Add the `rules.MTM2TotalForceMoment` rule.
+* Add the `rules.MTM2ForceError` rule.
 
 v1.18.1
 -------